move ActionsPanel, restore menuitem roles to primarynav
