+++ src/components/ActionsPanel/index.tsx
@@ -1,0 +1,2 @@
+import ActionsPanel from './ActionsPanel';
+export default ActionsPanel;
